Fix warnings (Clang)

+++ test/DynArrayTests.cpp
@@ -13,23 +13,23 @@ TEST_CASE("Create and destroy", "[RERR_DynArray]") {
 }
 
 TEST_CASE("Back insert and erase", "[RERR_DynArray]") {
-    RERR_DynArrayPtr a = RERR_DynArray_Create(sizeof(int));
+    RERR_DynArrayPtr a = RERR_DynArray_Create(sizeof(size_t));
     REQUIRE(a != nullptr);
 
-    for (int i = 0; i < 10; ++i) {
+    for (size_t i = 0; i < 10; ++i) {
         REQUIRE(RERR_DynArray_GetSize(a) == i);
         auto p = RERR_DynArray_Insert(a, RERR_DynArray_End(a));
-        *static_cast<int *>(p) = i + 42;
-        REQUIRE(*static_cast<int *>(RERR_DynArray_Back(a)) == i + 42);
-        REQUIRE(*static_cast<int *>(RERR_DynArray_Front(a)) == 42);
+        *static_cast<size_t *>(p) = i + 42;
+        REQUIRE(*static_cast<size_t *>(RERR_DynArray_Back(a)) == i + 42);
+        REQUIRE(*static_cast<size_t *>(RERR_DynArray_Front(a)) == 42);
     }
 
-    for (int i = 0; i < 10; ++i) {
+    for (size_t i = 0; i < 10; ++i) {
         auto p = RERR_DynArray_At(a, i);
-        REQUIRE(*static_cast<int *>(p) == i + 42);
+        REQUIRE(*static_cast<size_t *>(p) == i + 42);
     }
 
-    for (int i = 0; i < 10; ++i) {
+    for (size_t i = 0; i < 10; ++i) {
         REQUIRE(RERR_DynArray_GetSize(a) == 10 - i);
         auto p = RERR_DynArray_Erase(a, RERR_DynArray_Back(a));
         REQUIRE(p == RERR_DynArray_End(a));
@@ -40,23 +40,23 @@ TEST_CASE("Back insert and erase", "[RERR_DynArray]") {
 }
 
 TEST_CASE("Front insert and erase", "[RERR_DynArray]") {
-    RERR_DynArrayPtr a = RERR_DynArray_Create(sizeof(int));
+    RERR_DynArrayPtr a = RERR_DynArray_Create(sizeof(size_t));
     REQUIRE(a != nullptr);
 
-    for (int i = 0; i < 10; ++i) {
+    for (size_t i = 0; i < 10; ++i) {
         REQUIRE(RERR_DynArray_GetSize(a) == i);
         auto p = RERR_DynArray_Insert(a, RERR_DynArray_Begin(a));
         *static_cast<int *>(p) = i + 42;
-        REQUIRE(*static_cast<int *>(RERR_DynArray_Front(a)) == i + 42);
-        REQUIRE(*static_cast<int *>(RERR_DynArray_Back(a)) == 42);
+        REQUIRE(*static_cast<size_t *>(RERR_DynArray_Front(a)) == i + 42);
+        REQUIRE(*static_cast<size_t *>(RERR_DynArray_Back(a)) == 42);
     }
 
-    for (int i = 0; i < 10; ++i) {
+    for (size_t i = 0; i < 10; ++i) {
         auto p = RERR_DynArray_At(a, i);
-        REQUIRE(*static_cast<int *>(p) == (10 - i - 1) + 42);
+        REQUIRE(*static_cast<size_t *>(p) == (10 - i - 1) + 42);
     }
 
-    for (int i = 0; i < 10; ++i) {
+    for (size_t i = 0; i < 10; ++i) {
         REQUIRE(RERR_DynArray_GetSize(a) == 10 - i);
         auto p = RERR_DynArray_Erase(a, RERR_DynArray_Front(a));
         REQUIRE(p == RERR_DynArray_Begin(a));
